refactor GUI controls to seperate souce file
